refactor: group files features
